backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv
httpx
numpy
psycopg2-binary
pymupdf
python-docx
openpyxl
chardet
watchfiles
